change
git-svn-id: file:///var/www/svn/mxtce-sw@217 52860be4-34ab-47e5-b986-9ec2afffd7f9

+++ java/tcePCE/src/main/java/net/es/oscars/pce/tce/ReplyBagSegmentContent.java
@@ -2,15 +2,15 @@ package net.es.oscars.pce.tce;
 
 public class ReplyBagSegmentContent {
 	
-	protected int bandwidth;
+	protected long bandwidth;
 	protected int startTime;
 	protected int endTime;
 	
-	public int getBandwidth(){
+	public long getBandwidth(){
 		return this.bandwidth;
 	}
 	
-	public void setBandwidth(int value){
+	public void setBandwidth(long value){
 		this.bandwidth = value;
 	}
